\docType{data}
\name{earthquake}
\alias{earthquake}
\title{The Earthquake Data Set}
\format{A 45 by 2 matrix}
\description{
  The earthquake dataset contains...
}
\details{
  Details on ...
}
\examples{
## Load the dataset into R session:
data(earthquake)
}
\references{
  reference here
}
\keyword{datasets}

