% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.nb.reg.1.R
\name{fit.nb.reg.1}
\alias{fit.nb.reg.1}
\title{Implement genewise dispersion approach (report dispersion and fitted means)}
\usage{
fit.nb.reg.1(y, s, x, beta0 = rep(NA, dim(x)[2]), kappa = NA)
}
\arguments{
\item{kappa}{}
}
\value{
phi.hat and mu.hat
}
\author{
Yanming Di Gu Mi
}
