% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genewise.R
\name{genewise}
\alias{genewise}
\title{Implement genewise dispersion approach (but without any tests)}
\usage{
genewise(nb.data, grp.ids, grp1, grp2, R = 100)
}
\arguments{
\item{R}{}
}
\value{
a data frame where each row is from hoa.1u() function output
}
\author{
Yanming Di Gu Mi
}
