% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model.edgeR.tagtrd}
\alias{model.edgeR.tagtrd}
\title{Modeling NB Tagwise-Trended Dispersion with the Adjusted Profile Likelihood
Estimator (APLE) on Original and Simulated Datasets}
\usage{
model.edgeR.tagtrd(counts, x, lib.sizes=colSums(counts), prior.df=prior.df, min.n=min.n, method=method)
}
\arguments{
\item{counts}{an m-by-n count matrix of non-negative integers. For a typical
RNA-Seq experiment, this is the read counts with m genes and n samples.}

\item{x}{an n-by-p design matrix.}

\item{lib.sizes}{library sizes of an RNA-Seq experiment. Default is the column
sums of the \code{counts} matrix.}

\item{prior.df}{prior degrees of freedom to control the level of shrinkage (default is 10). See
\code{\link{estimateGLMTagwiseDisp}} for more details.
This argument is controlled by a higher level function \code{\link{nb.gof.m}}.}

\item{min.n}{minimim number of genes in a bin. Default is 100. See \code{\link{dispBinTrend}}
for details (lower-level function of \code{\link{estimateGLMTrendedDisp}}).}

\item{method}{method for estimating the trended dispersion, including "auto", "bin.spline", "bin.loess", "power" and "spline".
If NULL, then the "auto" method. Normally the number of tags analyzed is greater than 200, so the "bin.spline" method is used which
calls the \code{\link{dispBinTrend}} function. See \code{\link{estimateGLMTrendedDisp}} for more details.}
}
\value{
A list of quantities to be used in the main \code{\link{nb.gof.m}} function.
}
\description{
This function fits an NB regression model with
tagwise dispersions using the adjusted profile likelihood estimator (APLE). In \code{edgeR},
this tagwise dispersion model lies between two extreme scenarios: entirely
individual \eqn{\phi_i} for each gene and entirely shared values (i.e. common dispersion).
The function \code{estimateGLMTagwiseDisp} in \code{edgeR} shrinks the dispersion towards
a common dispersion or a global dispersion trend. In our implementation, the function
\code{model.edgeR.tagtrd}
shrinks the tagwise dispersions towards the global dispersion trend. See details below.
The output of this function will be passed to the main GOF function \code{\link{nb.gof.m}}.
}
\details{
In this tagwise-trended dispersion model, the dispersion parameter \eqn{\phi_i} is
estimated by maximizing a penalized log-likelihood \eqn{APL_g(\phi_g)} plus a weighted shared
log-likelihood. The weight denoted by \eqn{G_0} controls the level of shrinkage applied to
purely genewise dispersions towards a common dispersion. In the paper of
McCarthy et al. (2012),
\eqn{G_0=20/df} performs well over real RNA-Seq datasets. Here the numerator 20 is the prior
degrees of freedom, which can be specified by the argument \code{prior.df} in the lower-level
function \code{\link{dispCoxReidInterpolateTagwise}} in \code{edgeR}. The denominator "df" is
the residual degrees of freedom (number of libraries minus the number of distinct treatment
groups). We leave the default value of \code{prior.df} at 10. See the
\code{\link{estimateGLMTrendedDisp}}, \code{\link{estimateGLMTagwiseDisp}},
\code{\link{dispCoxReidInterpolateTagwise}}, lower-level functions for the \code{edgeR}
trended dispersion models, and \code{\link{glmFit}} functions in the
\code{edgeR} package for more details.
}
\author{
Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel Schafer
}
\references{
See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}
\seealso{
\code{\link{model.edgeR.genewise}} for the non-shrinkage, entirely genewise model,
and \code{\link{model.edgeR.tagcom}} for the tagwise model that shrinks towards a common
dispersion.
}

