% Generated by roxygen2 (4.0.1): do not edit by hand
\name{model.edgeR.trended}
\alias{model.edgeR.trended}
\title{Modeling NB Trended (Non-parametric) Dispersion with the Adjusted Profile
Likelihood estimator (APLE) on Original and Simulated Datasets}
\usage{
model.edgeR.trended(counts, x, lib.sizes=colSums(counts), min.n=min.n, method=method)
}
\arguments{
\item{counts}{an m-by-n count matrix of non-negative integers. For a typical
RNA-Seq experiment, this is the read counts with m genes and n samples.}

\item{x}{an n-by-p design matrix.}

\item{lib.sizes}{library sizes of an RNA-Seq experiment. Default is the column
sums of the \code{counts} matrix.}

\item{min.n}{minimim number of genes in a bin. Default is 100. See \code{\link{dispBinTrend}}
for details (lower-level function of \code{\link{estimateGLMTrendedDisp}}).}

\item{method}{method for estimating the trended dispersion, including "auto", "bin.spline", "bin.loess", "power" and "spline".
If NULL, then the "auto" method. Normally the number of tags analyzed is greater than 200, so the "bin.spline" method is used which
calls the \code{\link{dispBinTrend}} function. See \code{\link{estimateGLMTrendedDisp}} for more details.}
}
\value{
A list of quantities to be used in the main \code{\link{nb.gof.m}} function.
}
\description{
This function fits an NB regression model with
trended (non-parametric) dispersions using the adjusted profile likelihood estimator.
In \code{edgeR}, this function assumes the dispersion \eqn{\phi_i} satisfies
\eqn{\phi_i=s(\bar{\mu}_{i\cdot})}, where \eqn{s(\cdot)} is a smooth function of
each gene's average read counts across samples. A variety of non-parametric
approaches can be used by fitting loess or spline curves on binned genes, or
using locally weighted APL. See details below. The output of this function will
be passed to the main GOF function \code{\link{nb.gof.m}}.
}
\details{
In this trended non-parametric model, \eqn{\phi_{ij}} is estimated in a
first step as a smooth function of
\eqn{\log(\hat{\phi}_{ij})} on \eqn{\log(\hat{\mu}_{ij})}, and then treated as known
in the second step of regression coefficient inference. See the
\code{\link{estimateGLMTrendedDisp}} and
\code{\link{glmFit}} functions in the \code{\link{edgeR}} package
for more information.
}
\author{
Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel Schafer
}
\references{
See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}

