% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.genewise.R
\name{model.genewise}
\alias{model.genewise}
\title{Modeling NB Genewise Dispersion Using NBPSeq}
\usage{
model.genewise(counts, x)
}
\arguments{
\item{counts}{an m-by-n count matrix of non-negative integers. For a typical
RNA-Seq experiment, this is the read counts with m genes and n samples.}

\item{x}{an n-by-p design matrix.}
}
\value{
A list of quantities to be used in the main \code{\link{nb.gof.m}} function.
}
\description{
This function fits an NB regression model with
genewise dispersions using the adjusted profile likelihood estimator. See details below. The output of this function 
will be passed to the main GOF function \code{\link{nb.gof.m}}.
}
\details{
details here (HOA)
}
\author{
Gu Mi <neo.migu@gmail.com>, Yanming Di, Daniel Schafer
}
\references{
See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}

