% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model.nb2.v}
\alias{model.nb2.v}
\title{Modeling NB2 Regression Model with Maximum Likelihood (ML) on Original and Simulated Datasets}
\usage{
model.nb2.v(y, x, lib.sizes=NULL)
}
\arguments{
\item{y}{an n-by-1 vector of non-negative integers. For a typical RNA-Seq experiment,
this may represent the read counts for a single gene across n samples.}

\item{x}{an n-by-p design matrix. If an intercept is desired in the model, you need to specify
the first column of \code{x} as a vector of 1.}

\item{lib.sizes}{library sizes of a RNA-Seq experiment. Default is 1 for all samples.}
}
\value{
A list of quantities to be used in the main \code{\link{nb.gof.v}} function.
}
\description{
This function is designed to fit an NB2 regression model. The output of
this function will be passed to the main GOF function \code{\link{nb.gof.v}}.
}
\details{
The \code{nb.regression.1} function is used for NB2 model fitting with MLE.
}
\author{
Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel Schafer
}
\references{
See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}

