% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.nbp.v.R
\name{model.nbp.v}
\alias{model.nbp.v}
\title{Modeling NBP Regression Model with Maximum Likelihood (ML) or Adjusted Profile
Likelihood (APL) on Original and Simulated Datasets}
\usage{
model.nbp.v(y, x, lib.sizes=NULL, method="ML")
}
\arguments{
\item{y}{an n-by-1 vector of non-negative integers. For a typical RNA-Seq experiment, 
this may represent the read counts for a single gene across n samples.}

\item{x}{an n-by-p design matrix. If an intercept is desired in the model, you need to specify
the first column of \code{x} as a vector of 1.}

\item{lib.sizes}{library sizes of a RNA-Seq experiment. Default is 1 for all samples.}

\item{method}{estimation method for NBP model fit. Either maximum likelihood (\code{ML})
or adjusted profile likelihood (\code{APL}). Default is \code{ML}.}
}
\value{
A list of quantities to be used in the main \code{\link{nb.gof.v}} function.
}
\description{
This function is designed to fit an NBP regression model. The output of
this function will be passed to the main GOF function \code{\link{nb.gof.v}}.
}
\details{
The \code{glm.nbp.1.MLE} function is used for NBP model fitting with MLE.
The \code{glm.nbp.1} function is used for NBP model fitting with APLE.
}
\references{
Di Y, Schafer DW, Cumbie JS, and Chang JH (2011): "The NBP Negative Binomial
Model for Assessing Differential Gene Expression from RNA-Seq", \emph{Statistical 
Applications in Genetics and Molecular Biology}, 10 (1).

See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}
\author{
Gu Mi <neo.migu@gmail.com>, Yanming Di, Daniel Schafer
}
