% Generated by roxygen2 (4.0.1): do not edit by hand
\name{model.nbq.m}
\alias{model.nbq.m}
\title{Modeling NBQ Genewise Dispersion with the Maximum Ajdusted Profile Likelihood (MAPL) Estimator
on Original and Simulated Datasets}
\usage{
model.nbq.m(counts, x, lib.sizes=colSums(counts), method=method)
}
\arguments{
\item{counts}{an m-by-n count matrix of non-negative integers. For a typical
RNA-Seq experiment, this is the read counts with m genes and n samples.}

\item{x}{an n-by-p design matrix.}

\item{lib.sizes}{library sizes of an RNA-Seq experiment. Default is the column
sums of the \code{counts} matrix.}

\item{method}{method for estimating dispersions.}
}
\value{
A list of quantities to be used in the main \code{\link{nb.gof.m}} function.
}
\description{
This function fits an NBQ dispersion model where the dispersion parameter
is modeled as a linear function of the relative means, plus a quadratic term. See details below.
The output of this function will be passed to the main GOF function \code{\link{nb.gof.m}}.
}
\details{
Under the NB model, the mean-variance relationship of a single read count
satisfies \eqn{\sigma_{ij}^2 = \mu_{ij} + \phi_{ij} \mu_{ij}^2}. For applying the NBP
model to RNA-Seq data, we consider a quadratic model assuming a
parametric dispersion model
\eqn{ \phi_{ij} = \alpha_0 + \alpha_1 \log(\pi_{ij}) + \alpha_2 (\log(\pi_{ij}))^2},
where \eqn{\pi_{ij} = \mu_{ij}/(N_j R_j)} is the relative mean frequency after
normalization. The parameters \eqn{(\alpha_0, \alpha_1, \alpha_2)} in this dispersion model
are estimated by maximizing the adjusted profile likelihood. See the
\code{\link{estimate.dispersion}} function in the \code{NBPSeq} package
for more information.
}
\author{
Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel Schafer
}
\references{
Di Y, Schafer DW, Cumbie JS, and Chang JH (2011): "The NBP Negative Binomial
Model for Assessing Differential Gene Expression from RNA-Seq", \emph{Statistical
Applications in Genetics and Molecular Biology}, 10 (1).

See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}

