% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model.poi.v}
\alias{model.poi.v}
\title{Modeling Poisson Regression Model on Original and Simulated Datasets}
\usage{
model.poi.v(y, x, lib.sizes=NULL)
}
\arguments{
\item{y}{an n-by-1 vector of non-negative integers. For a typical RNA-Seq experiment,
this may represent the read counts for a single gene across n samples.}

\item{x}{an n-by-p design matrix. For Poisson model fitting, we used the \link{glm} function,
so if an intercept is desired, there is no need to include the first column of 1.}

\item{lib.sizes}{library sizes of a RNA-Seq experiment. Default is 0 for all samples.}
}
\value{
A list of quantities to be used in the main \code{\link{nb.gof.v}} function.
}
\description{
This function is designed to fit a Poisson regression model. The output of
this function will be passed to the main GOF function.
}
\details{
The \code{glm} function with \code{family=poisson} is used for Poisson model fitting.
}
\author{
Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel Schafer
}
\references{
See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}

