% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sort.vec}
\alias{sort.vec}
\title{Group-wise sort of residual matrix in multivariate case}
\usage{
\method{sort}{vec}(x, grp.ids)
}
\description{
Group-wise sort of residual matrix in multivariate case.
}
\author{
Gu Mi <neo.migu@gmail.com>, Yanming Di, Daniel Schafer
}
\keyword{internal}

