% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPPlot.R
\name{EPPlot}
\alias{EPPlot}
\title{Empirical Probability Plots of Negative Binomial Regression Models}
\usage{
EPPlot(gofv.obj, envelope=0.95, data.note=NULL)
}
\arguments{
\item{gofv.obj}{an object of class "gofv" from the \code{\link{nb.gof.v}} output.}

\item{envelope}{the level of prediction band when drawing EPP.}

\item{data.note}{user-specified notes on lower right part of the plot.}
}
\value{
A ggplot object of empirical probability plot.
}
\description{
This function makes empirical probability plots (EPP) based on the goodness-of-fit test results (an "gofv" object) from
\code{\link{nb.gof.v}}.
}
\author{
Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel Schafer
}
\references{
See \url{https://github.com/gu-mi/NBGOF/wiki/} for more details.
}
\seealso{
\code{\link{nb.gof.v}} for simulated data examples.
}

