% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit.nb.reg.1}
\alias{fit.nb.reg.1}
\title{Implement genewise dispersion approach (report dispersion and fitted means)}
\usage{
fit.nb.reg.1(y, s, x, beta0 = rep(NA, dim(x)[2]), kappa = NA)
}
\arguments{
\item{y}{}

\item{s}{}

\item{x}{}

\item{beta0}{}

\item{kappa}{}
}
\value{
phi.hat and mu.hat
}
\description{
Implement genewise dispersion approach (report dispersion and fitted means)
}
\author{
Yanming Di Gu Mi
}

