% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genewise}
\alias{genewise}
\title{Implement genewise dispersion approach (but without any tests)}
\usage{
genewise(nb.data, grp.ids, grp1, grp2, R = 100)
}
\arguments{
\item{nb.data}{}

\item{grp.ids}{}

\item{grp1}{}

\item{grp2}{}

\item{R}{}
}
\value{
a data frame where each row is from hoa.1u() function output
}
\description{
Implement genewise dispersion approach (but without any tests)
}
\author{
Yanming Di Gu Mi
}

