\name{sort.vec}
\alias{sort.vec}
\title{Group-wise sort of residual matrix in multivariate case}
\usage{
  sort.vec(x, grp.ids)
}
\description{
  Group-wise sort of residual matrix in multivariate case.
}
\author{
  Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel
  Schafer
}
\keyword{internal}

