% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sort.vec}
\alias{sort.vec}
\title{Group-wise sort of residual matrix in multivariate case}
\usage{
\method{sort}{vec}(x, grp.ids)
}
\description{
Group-wise sort of residual matrix in multivariate case.
}
\author{
Gu Mi <mig@stat.oregonstate.edu>, Yanming Di, Daniel Schafer
}
\keyword{internal}

